requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.0.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.12
decorator==5.1.1
dnspython==2.4.0
entrypoints==0.4
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.235
langsmith==0.0.7
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
mdurl==0.1.2
mpmath==1.3.0
msal==1.25.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
pinecone-client==2.2.2
portalocker==2.8.2
protobuf==3.20.3
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
PyMuPDF==1.22.5
pypdf==3.12.2
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
streamlit==1.28.1
streamlit-chat==0.1.1
streamlit-chat-media==0.0.4
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2